+++ jsonschema-module-javax-validation/src/main/java/com/github/victools/jsonschema/module/javax/validation/JavaxValidationModule.java
@@ -29,6 +29,8 @@ import java.util.Collections;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.function.Function;
+import java.util.function.Predicate;
+import javax.validation.Constraint;
 import javax.validation.constraints.DecimalMax;
 import javax.validation.constraints.DecimalMin;
 import javax.validation.constraints.Email;
@@ -57,6 +59,8 @@ import javax.validation.constraints.Size;
  * </ul>
  */
 public class JavaxValidationModule implements Module {
+
+    private static final Predicate<Annotation> CONSTRAINT_CHECK = annotation -> annotation.annotationType().isAnnotationPresent(Constraint.class);
 
     private final Set<JavaxValidationOption> options;
     private Set<Class<?>> validationGroups;
@@ -144,11 +148,11 @@ public class JavaxValidationModule implements Module {
      */
     protected <A extends Annotation> A getAnnotationFromFieldOrGetter(MemberScope<?, ?> member, Class<A> annotationClass,
             Function<A, Class<?>[]> validationGroupsLookup) {
-        A containerItemAnnotation = member.getContainerItemAnnotationConsideringFieldAndGetterIfSupported(annotationClass);
+        A containerItemAnnotation = member.getContainerItemAnnotationConsideringFieldAndGetterIfSupported(annotationClass, CONSTRAINT_CHECK);
         if (this.shouldConsiderAnnotation(containerItemAnnotation, validationGroupsLookup)) {
             return containerItemAnnotation;
         }
-        A annotation = member.getAnnotationConsideringFieldAndGetterIfSupported(annotationClass);
+        A annotation = member.getAnnotationConsideringFieldAndGetterIfSupported(annotationClass, CONSTRAINT_CHECK);
         if (this.shouldConsiderAnnotation(annotation, validationGroupsLookup)) {
             return annotation;
         }